backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
aiosqlite>=0.20.0
fastapi-users>=14.0.0
fastapi-users-db-sqlalchemy>=7.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
